dialects: (qu) rename from qubit to qu
