+++ tests/filecheck/dialects/qubit.mlir
@@ -1,10 +1,10 @@
 // RUN: QUOPT_ROUNDTRIP
 // RUN: QUOPT_GENERIC_ROUNDTRIP
 
-// CHECK: %q = qubit.alloc
-// CHECK-GENERIC: %q = "qubit.alloc"() <{alloc = #qubit.zero}> : () -> !qubit.bit
-%q = qubit.alloc
+// CHECK: %q = qu.alloc
+// CHECK-GENERIC: %q = "qu.alloc"() <{alloc = #qu.zero}> : () -> !qu.bit
+%q = qu.alloc
 
-// CHECK: %q2 = qubit.alloc<#qubit.plus>
-// CHECK-GENERIC: %q2 = "qubit.alloc"() <{alloc = #qubit.plus}> : () -> !qubit.bit
-%q2 = qubit.alloc<#qubit.plus>
+// CHECK: %q2 = qu.alloc<#qu.plus>
+// CHECK-GENERIC: %q2 = "qu.alloc"() <{alloc = #qu.plus}> : () -> !qu.bit
+%q2 = qu.alloc<#qu.plus>
