+++ tests/filecheck/transforms/convert_qssa_to_qref.mlir
@@ -1,8 +1,8 @@
 // RUN: quopt -p convert-qssa-to-qref %s | filecheck %s
 // RUN: quopt -p convert-qssa-to-qref,convert-qref-to-qssa %s | filecheck %s --check-prefix=CHECK-ROUNDTRIP
 
-%q0 = qubit.alloc
-%q1 = qubit.alloc
+%q0 = qu.alloc
+%q1 = qu.alloc
 %q2 = qssa.gate<#gate.h> %q0
 %q3 = qssa.gate<#gate.rz<0.5pi>> %q1
 %q4, %q5 = qssa.gate<#gate.cx> %q2, %q3
@@ -12,8 +12,8 @@
 %m = measurement.constant #measurement.comp_basis
 %1 = qssa.dyn_measure<%m> %q6
 
-// CHECK:      %q0 = qubit.alloc
-// CHECK-NEXT: %q1 = qubit.alloc
+// CHECK:      %q0 = qu.alloc
+// CHECK-NEXT: %q1 = qu.alloc
 // CHECK-NEXT: qref.gate<#gate.h> %q0
 // CHECK-NEXT: qref.gate<#gate.rz<0.5pi>> %q1
 // CHECK-NEXT: qref.gate<#gate.cx> %q0, %q1
@@ -23,8 +23,8 @@
 // CHECK-NEXT: %m = measurement.constant #measurement.comp_basis
 // CHECK-NEXT: %{{.*}} = qref.dyn_measure<%m> %q1
 
-// CHECK-ROUNDTRIP:      %q0 = qubit.alloc
-// CHECK-ROUNDTRIP-NEXT: %q1 = qubit.alloc
+// CHECK-ROUNDTRIP:      %q0 = qu.alloc
+// CHECK-ROUNDTRIP-NEXT: %q1 = qu.alloc
 // CHECK-ROUNDTRIP-NEXT: %q0_1 = qssa.gate<#gate.h> %q0
 // CHECK-ROUNDTRIP-NEXT: %q1_1 = qssa.gate<#gate.rz<0.5pi>> %q1
 // CHECK-ROUNDTRIP-NEXT: %q0_2, %q1_2 = qssa.gate<#gate.cx> %q0_1, %q1_1
