+++ tests/filecheck/transforms/convert_to_mbqc.mlir
@@ -1,10 +1,10 @@
 // RUN: quopt %s -p convert-to-mbqc | filecheck %s
 
-// CHECK:      func.func @rotation(%phi : !angle.type, %theta : !angle.type, %lambda : !angle.type, %q1 : !qubit.bit) -> !qubit.bit {
-// CHECK-NEXT:   %q1_1 = qubit.alloc<#qubit.plus>
-// CHECK-NEXT:   %q1_2 = qubit.alloc<#qubit.plus>
-// CHECK-NEXT:   %q1_3 = qubit.alloc<#qubit.plus>
-// CHECK-NEXT:   %q1_4 = qubit.alloc<#qubit.plus>
+// CHECK:      func.func @rotation(%phi : !angle.type, %theta : !angle.type, %lambda : !angle.type, %q1 : !qu.bit) -> !qu.bit {
+// CHECK-NEXT:   %q1_1 = qu.alloc<#qu.plus>
+// CHECK-NEXT:   %q1_2 = qu.alloc<#qu.plus>
+// CHECK-NEXT:   %q1_3 = qu.alloc<#qu.plus>
+// CHECK-NEXT:   %q1_4 = qu.alloc<#qu.plus>
 // CHECK-NEXT:   %q1_5, %q1_6 = qssa.gate<#gate.cz> %q1, %q1_1
 // CHECK-NEXT:   %0, %1 = qssa.gate<#gate.cz> %q1_6, %q1_2
 // CHECK-NEXT:   %2, %3 = qssa.gate<#gate.cz> %1, %q1_3
@@ -27,9 +27,9 @@
 // CHECK-NEXT:   %q1_21 = arith.xori %q1_20, %q1_11 : i1
 // CHECK-NEXT:   %6 = gate.xz %q1_21, %q1_16
 // CHECK-NEXT:   %q1_22 = qssa.dyn_gate<%6> %5
-// CHECK-NEXT:   func.return %q1_22 : !qubit.bit
+// CHECK-NEXT:   func.return %q1_22 : !qu.bit
 // CHECK-NEXT: }
-func.func @rotation(%phi: !angle.type, %theta: !angle.type, %lambda: !angle.type, %q1: !qubit.bit) -> !qubit.bit {
+func.func @rotation(%phi: !angle.type, %theta: !angle.type, %lambda: !angle.type, %q1: !qu.bit) -> !qu.bit {
   %q1_1 = qssa.gate<#gate.j<0>> %q1
   %g1 = gate.dyn_j<%lambda>
   %q1_2 = qssa.dyn_gate<%g1> %q1_1
@@ -37,5 +37,5 @@ func.func @rotation(%phi: !angle.type, %theta: !angle.type, %lambda: !angle.type
   %q1_3 = qssa.dyn_gate<%g2> %q1_2
   %g3 = gate.dyn_j<%phi>
   %q1_4 = qssa.dyn_gate<%g3> %q1_3
-  func.return %q1_4 : !qubit.bit
+  func.return %q1_4 : !qu.bit
 }
